Do not add empty remarks item in target Docs xml.
There are certain API elements (types or members) that do not have a remarks item by default in Docs xmls. Currently, the tool automatically adds it if it's missing, but it shouldn't.

The logic has been modified so that the remarks item only gets added if there is a remarks section in the triple slash xml for that same API element.